requirements.txt
asgiref==3.3.1
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-heroku==0.3.1
django-nine==0.2.4
django-logging-json==1.15
djangorestframework==3.12.2
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
gunicorn==20.0.4
numpy==1.19.4
pandas==1.1.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
